move useLogVirtualizer
